+++ processor/resourcedetectionprocessor/config.go
@@ -21,6 +21,7 @@ import (
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/resourcedetectionprocessor/internal"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/resourcedetectionprocessor/internal/aws/ec2"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/resourcedetectionprocessor/internal/consul"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/resourcedetectionprocessor/internal/system"
 )
 
@@ -45,6 +46,10 @@ type Config struct {
 type DetectorConfig struct {
 	// EC2Config contains user-specified configurations for the EC2 detector
 	EC2Config ec2.Config `mapstructure:"ec2"`
+
+	// ConsulConfig contains user-specified configurations for the Consul detector
+	ConsulConfig consul.Config `mapstructure:"consul"`
+
 	// SystemConfig contains user-specified configurations for the System detector
 	SystemConfig system.Config `mapstructure:"system"`
 }
@@ -53,6 +58,8 @@ func (d *DetectorConfig) GetConfigFromType(detectorType internal.DetectorType) i
 	switch detectorType {
 	case ec2.TypeStr:
 		return d.EC2Config
+	case consul.TypeStr:
+		return d.ConsulConfig
 	case system.TypeStr:
 		return d.SystemConfig
 	default: